# integration-test/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sample E-commerce Application
  description: |-
    This is a ***Sample Ecommerce API*** based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    
    Some useful links:
    - [The Cometbid-SFI repository](https://github.com/cometbid-sfi/swagger-ecommerce)
    - [The source API definition for the Ecommerce appliation](https://github.com/cometbid-sfi/swagger-ecommerce/blob/master/src/main/resources/openapi.yaml)
  termsOfService: https://github.com/cometbid-sfi/swagger-ecommerce/blob/main/LICENSE
  contact:
    name: The Cometbid Software Foundation Inc.
    url: http://www.cometbid.org
    email: support@cometbid.org
  license:
    name: MIT License
    url: https://github.com/cometbid-sfi/swagger-ecommerce/blob/main/LICENSE
  version: 1.0.0
...
servers:
  - url: https://ecommerce.swagger.io/api/v2
...
paths:
...
  /api/v1/products/{id}/uploadImages:
    post:
      tags:
        - product
      summary: upload images
      description: ''
      operationId: uploadFile
      parameters:
        - name: id
          in: path
          description: ID of the product to update
          required: true
          schema:
            type: string
      requestBody:
        description: Multiple files to upload
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductImages'
      responses:
        '200':
          description: For successful upload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppResponse'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppResponse' 
        '400':
          description: Invalid file format or content type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/AppResponse'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        path:
          description: REST endpoint URI of the Request
          type: string
        message:
          description: short description (of the error)
          type: string
        method:
          description: Request method type 'GET, POST, PUT, DELETE' etc.
          example: 'GET'
          type: string
        status:
          description: HttpStatus of the (error) response
          type: string
        code:
          description: Unique application (error) code
          type: string
        traceId:
          description: unique id used to track error in a log
          type: string
        timestamp:
          description: date and time the error occured
          type: string
          format: date-time
        debugMessage:
          description: detail error message (of the error) to aid debugging
          type: string
        statusCode:
          description: status code of the error response
          type: string
        errorDetails:
          description: a breakdown of the error
          type: array
          items:
            $ref: '#/components/schemas/ApiSubError'
      xml:
        name: ApiResponse   
    ApiSubError:         
      type: object
      properties:
        object:
          description: object the caused the error
          type: string
        code:
          description: unique code of the error
          type: string
        error:
          description: error message 
          type: string
      xml:
        name: ApiSubError
    AppResponse:
      type: object
      properties:
        success: 
          description: Flag indicating whether the request succeeded or failed
          default: false
          type: boolean
        message: 
          description: simple message about the outcome of the request
          type: string  
        metadata: 
          $ref: '#/components/schemas/AppResponseMetadata'
        response: 
          $ref: '#/components/schemas/ApiResponse'
      xml:
        name: AppResponse   
    AppResponseMetadata:
      type: object
      properties: 
        apiVersion:
          description: Current version of the Api
          type: string
          example: 1.1.0
        sendReportUri:
          description: URL to with a form to send error report
          type: string
        moreInfoUrl:
          description: URL link to frequently asked questions
          type: string
        technical:
          description: Email address of the technical team
          type: string
      xml:
        name: AppResponseMetadata    
    Image:
      type: object
      properties:
        metadata:
          description: Image metadata
          type: string
        file:
          description: image binaries
          type: string
          format: binary
    ProductImages:
      type: object
      description: Images that will be associated with the Product 
      properties:
        length:
          type: integer
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Image'
      xml:
        name: ProductImages
...
  securitySchemes:
    store_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://ecommerce.swagger.io/oauth/authorize
          scopes:
            write:itemd: modify products in a store
            read:items: read products
    api_key:
      type: apiKey
      name: api_key
      in: header